Move loadFromSchema.tsx
